feat: create data validators and errors handlers

+++ src/modules/users/users.service.ts
@@ -3,7 +3,7 @@ import { Injectable } from '@nestjs/common';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import { UsersRepository } from './repositories/users.repository';
-
+import { NotFoundException } from '@nestjs/common';
 @Injectable()
 export class UsersService {
   constructor(private usersRepository: UsersRepository) {}
@@ -17,14 +17,26 @@ export class UsersService {
   }
 
   findOne(id: string) {
-    return this.usersRepository.findOne(id);
+    const findUser = this.usersRepository.findOne(id);
+    if (!findUser) {
+      throw new NotFoundException('user not found');
+    }
+    return findUser;
   }
 
   update(id: string, updateUserDto: UpdateUserDto) {
+    const findUser = this.usersRepository.findOne(id);
+    if (!findUser) {
+      throw new NotFoundException('user not found');
+    }
     return this.usersRepository.update(id, updateUserDto);
   }
 
   remove(id: string) {
+    const findUser = this.usersRepository.findOne(id);
+    if (!findUser) {
+      throw new NotFoundException('user not found');
+    }
     return this.usersRepository.delete(id);
   }
 }
